Re-organize the storage integration tests to group related behaviors in traits.
